fix init config file
