+++ config/config.go
@@ -6,8 +6,8 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
+	"path/filepath"
 
-	"github.com/davecgh/go-spew/spew"
 	"github.com/fatih/color"
 	cp "github.com/nmrshll/go-cp"
 	gphotos "github.com/nmrshll/google-photos-api-client-go/lib-gphotos"
@@ -73,6 +73,15 @@ func Load() *Config {
 	return Cfg
 }
 
+var noConfigFoundMessage = color.CyanString(`
+No config file found at ~/.config/gphotos-uploader-cli/config.hjson
+Will now copy an example config file.
+Edit it by running:
+
+	nano ~/.config/gphotos-uploader-cli/config.hjson
+
+`)
+
 func loadConfigFile() *Config {
 	configPathAbsolute, err := cp.AbsolutePath(CONFIGPATH)
 	if err != nil {
@@ -81,25 +90,10 @@ func loadConfigFile() *Config {
 
 	// if no config file copy the default example and exit
 	if !fileshandling.IsFile(configPathAbsolute) {
-		fmt.Println(color.CyanString(`
-No config file found at ~/.config/gphotos-uploader-cli/config.hjson
-Will now copy an example config file.
-Edit it by running:
-
-	nano ~/.config/gphotos-uploader-cli/config.hjson
-
-`,
-		))
-		spew.Dump(configPathAbsolute)
-		f, err := os.Create(configPathAbsolute)
-		if err != nil {
-			log.Fatal(err)
+		fmt.Println(noConfigFoundMessage)
+		if err := initConfigFile(); err != nil {
+			log.Fatal(stacktrace.Propagate(err, "failed initializing config file"))
 		}
-		_, err = f.WriteString(ExampleConfig)
-		if err != nil {
-			log.Fatal(err)
-		}
-
 		os.Exit(0)
 	}
 
@@ -117,7 +111,34 @@ Edit it by running:
 	return config
 }
 
-const ExampleConfig = `{
+func initConfigFile() error {
+	configPathAbsolute, err := cp.AbsolutePath(CONFIGPATH)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	dirname := filepath.Dir(configPathAbsolute)
+	if _, err := os.Stat(dirname); os.IsNotExist(err) {
+		os.Mkdir(dirname, 0755)
+	}
+
+	f, err := os.Open(configPathAbsolute)
+	if err != nil {
+		f, err = os.Create(configPathAbsolute)
+		if err != nil {
+			return err
+		}
+	}
+	defer f.Close()
+
+	_, err = f.WriteString(exampleConfig)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+const exampleConfig = `{
   APIAppCredentials: {
     ClientID:     "20637643488-1hvg8ev08r4tc16ca7j9oj3686lcf0el.apps.googleusercontent.com",
     ClientSecret: "0JyfLYw0kyDcJO-pGg5-rW_P",
