+++ go.mod
@@ -1,0 +1,28 @@
+module github.com/nmrshll/gphotos-uploader-cli
+
+require (
+	cloud.google.com/go v0.34.0 // indirect
+	github.com/Nr90/imgsim v0.0.0-20180202144352-5caa057144b0
+	github.com/client9/xson v0.0.0-20180321172152-0e50cdfc08c0
+	github.com/davecgh/go-spew v1.1.1
+	github.com/fatih/color v1.7.0
+	github.com/godbus/dbus v4.1.0+incompatible // indirect
+	github.com/hashicorp/go-multierror v1.0.0 // indirect
+	github.com/mattn/go-colorable v0.0.9 // indirect
+	github.com/mattn/go-isatty v0.0.4 // indirect
+	github.com/nfnt/resize v0.0.0-20180221191011-83c6a9932646 // indirect
+	github.com/nmrshll/go-cp v0.0.0-20180115193924-61436d3b7cfa
+	github.com/nmrshll/google-photos-api-client-go v0.0.0-20180906090521-a94d2fbcfb40
+	github.com/nmrshll/oauth2-noserver v0.0.0-20180827223500-16b622b98a45 // indirect
+	github.com/nmrshll/rndm-go v0.0.0-20170430161430-8da3024e53de // indirect
+	github.com/palantir/stacktrace v0.0.0-20161112013806-78658fd2d177
+	github.com/skratchdot/open-golang v0.0.0-20160302144031-75fb7ed4208c // indirect
+	github.com/spf13/cobra v0.0.3
+	github.com/spf13/pflag v1.0.3 // indirect
+	github.com/steakknife/hamming v0.0.0-20180906055917-c99c65617cd3
+	github.com/zalando/go-keyring v0.0.0-20180221093347-6d81c293b3fb
+	golang.org/x/net v0.0.0-20181207154023-610586996380 // indirect
+	golang.org/x/oauth2 v0.0.0-20181203162652-d668ce993890
+	google.golang.org/api v0.0.0-20181213000619-1327b224df06
+	gopkg.in/h2non/filetype.v1 v1.0.5
+)
